Beginnig of work with Game_Cards.
Added:
map to the The_Game class;
card_stock to the card.h;

Mechanics for the Cards_Storage created:
1) Basis are stored as map in the client's constructor.
It receives the Basis ID wich to use (in future)
2) Vector is used to store real stack on the server side.
3) Client receives only one-by-one numbers of cards coming into the Game
and filles up the stack of Player's cards and Folds of Doors and
Treasures. It also can compute how many cards left in main stocks, but
doesn't know thei order. (In fact it is only the server who knows)