doorType has been added.
The strategy is to use children classes for all the cards,
but it looks like that different massives (maps) are necessary to store
all their info.
So there will be the large "main_map" where all the cards are stored;
And 6 separate maps regarding the card type (look at the doorType enum
to understand their meaning).
For the 6th map (special_mechanics) - they only store a pointer to a
specific game function that will be initiated. All these functions are
The_Game methods.